Disable strict-aliasing optimizations until they are fixed upstream
When building the ms-tpm-20-ref code with a modern gcc and -Wstrict-aliasing
there are a number of these warnings that need to be addressed upstream:

 dereferencing type-punned pointer will break strict-aliasing rules

-Wstrict-aliasing is flagging code that may break with compiler
optimizations that it unlocks. Since code in ms-tpm-20-ref is being
flagged by this, it should not be built with these optimizations
enabled for now. Setting -fno-strict-aliasing will both disable these
unsafe optimizations and suppress the warnings.

Signed-off-by: Jan Setje-Eilers <Jan.SetjeEilers@oracle.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@st.com>